+++ TP_1/src/biblioteca.h
@@ -16,4 +16,4 @@ double sumar(double a, double b);
 double restar(double a, double b);
 double multiplicar(double a, double b);
 double dividir(double a, double b);
-long factorial(unsigned n);
+long long factorial(unsigned n);
